+++ .dotnet/src/Custom/Chat/ChatCompletionOptions.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 
@@ -98,9 +99,7 @@ public partial class ChatCompletionOptions
     // CUSTOM:
     // - Renamed.
     // - Changed type to avoid BinaryData.
-    /// <summary>
-    /// Deprecated in favor of <see cref="ToolChoice"/>.
-    /// </summary>
+    [Obsolete($"This property is obsolete. Please use {nameof(ToolChoice)} instead.")]
     [CodeGenMember("FunctionCall")]
     public ChatFunctionChoice FunctionChoice { get; set; }
 
@@ -139,7 +138,7 @@ public partial class ChatCompletionOptions
     [CodeGenMember("User")]
     public string EndUserId { get; set; }
 
-    // CUSTOM: Added the Experimental attribute
+    // CUSTOM: Added the Experimental attribute.
     /// <summary>
     /// If specified, our system will make a best effort to sample deterministically, such that repeated requests with the same seed and parameters should return the same result.
     /// </summary>
@@ -172,4 +171,7 @@ public partial class ChatCompletionOptions
 
     [CodeGenMember("MaxCompletionTokens")]
     internal int? _maxCompletionTokens;
+    // CUSTOM: Added the Obsolete attribute.
+    [Obsolete($"This property is obsolete. Please use {nameof(Tools)} instead.")]
+    public IList<ChatFunction> Functions { get; }
 }